final commit
Description:
Design and implement a queue management application that assigns clients to queues in a way that minimizes their waiting time.

Queues are commonly used to model real-world systems. The primary objective of a queue is to provide a place for a "client" to wait before receiving a "service". Queue management systems aim to minimize the amount of time clients spend waiting before being served. One way to reduce waiting time is to add more servers (i.e., more queues, each associated with a processor). However, this approach increases the cost for the service provider.

The queue management application should simulate, over a defined simulation time (𝑡_simulation), a series of N clients arriving for service, entering one of Q queues, waiting, being served, and finally leaving the system.

All clients are generated at the start of the simulation. Each client is characterized by three parameters:
 - ID – a unique identifier (from 1 to N)
 - 𝑡_arrival – the simulation time at which the client is ready to enter a queue
 - 𝑡_service – the duration required to serve the client once at the front of the queue

The application tracks the total time spent by each client in the queues and calculates the average waiting time. When a client's 𝑡_arrival is less than or equal to the current simulation time (𝑡_arrival ≤ 𝑡_simulation), the client is added to the queue with the minimum total waiting time.

User Input (via GUI):
The following data should be entered by the user through the application’s graphical interface:
- Number of clients (N)
- Number of queues (Q)
- Simulation interval
- Minimum and maximum arrival time
- Minimum and maximum service time